Initialized Grunt, Moved main, Added .gitignore
Grunt is the tool being used to send API requests to screeps. This enables 3rd party uploading from any source through the API

Moved the main file to dist/main.js. This will be where all of the screeps scripts will be kept from now on

Added a .gitignore for node_modules